finder_app_be/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altgraph==0.17.4
and==66.0.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.6.2
CacheControl==0.14.3
cachetools==5.5.2
caugetch==0.0.1
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.2.0
click==8.1.8
clipboard==0.0.4
colorama==0.4.6
contourpy==1.1.0
crispy-bootstrap5==2024.2
cryptography==46.0.3
cv==1.0.0
cycler==0.11.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.2
django-cors-headers==4.8.0
django-crispy-forms==2.3
django-database-url==1.0.3
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.7
fastjsonschema==2.21.1
firebase_admin==7.1.0
flake8==7.3.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fonttools==4.41.1
frozenlist==1.4.1
getpass4==0.0.14.1
google==3.0.0
google-api-core==2.28.1
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==2.0.2
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.4
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyboard==0.13.5
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
mouse==0.7.1
msgpack==1.1.2
multidict==6.1.0
mutagen==1.47.0
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.25.1
openai==0.28.0
orjson==3.10.15
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
Pillow-PIL==0.1.dev0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.33.0
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.9.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.23.4
pyflakes==3.4.0
pygame==2.5.2
Pygments==2.15.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pymongo==4.4.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.12
python-socketio==5.11.4
PyTMX==3.32
pytube==15.0.0
pytube3==9.6.4
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==7.4.2
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.25.1
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.38.6
tinycss2==1.4.0
tk==0.1.0
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
unicorn==2.1.1
urllib3==2.0.5
uvicorn==0.31.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==2.3.6
whitenoise==6.9.0
win10toast==0.9
windows-curses==2.3.2
wsproto==1.2.0
yarg==0.1.9
yarl==1.13.1
zope.interface==6.1
